refactor: merge daily and weekly schedule handlers for students
